# app_web/docs/swagger_admin.yaml
openapi: 3.1.0
info:
  title: 6G IDS API - Administrator (Full Access)
  version: 1.0.0
  description: Full-access gateway API for IOTinel.
servers:
  - url: http://localhost:8010
...
paths:
...
  /train/runs/{run_id}:
    get:
      tags: [Training]
      summary: Run details
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run detail
  /train/promote/{run_id}:
    post:
      tags: [Training]
      summary: Promote champion
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Promotion
...